Clean up CLI
